完善insert.

+++ model-spring-suport/src/main/java/com/javaoffers/batis/modelhelper/fun/crud/WhereModifyFun.java
@@ -8,5 +8,5 @@ import com.javaoffers.batis.modelhelper.fun.GetterFun;
  * @Description:
  * @Auther: create by cmj on 2022/5/4 21:44
  */
-public interface WhereModifyFun<M,V> extends  WhereFun<M, GetterFun<M,V>,V, WhereModifyFun<M,V>> , ExecutOneFun<Integer> {
+public interface WhereModifyFun<M,V> extends  WhereFun<M, GetterFun<M,V>,V, WhereModifyFun<M,V>> , ExecutOneFun<Long> {
 }
